Use memory-sized types when possible

+++ programs/common.h
@@ -32,7 +32,7 @@
 
 #pragma once
 
-#define ARRAY_LEN(x) ((int)(sizeof(x) / sizeof(x[0])))
+#define ARRAY_LEN(x) (sizeof(x) / sizeof(x[0]))
 #define MAX(a, b) ((a) > (b) ? (a) : (b))
 
 typedef struct
